Adds AmazonProduct model.
The amazon-synchronizer service will, over time, migrate the `ASIN` and `Price` over to the `AmazonProducts` table.  Once the migration of data is complete, this service will be updated to pull the `Price`/`ASIN` from there via a `left outer join` (or simple navigation in EF).  Once this is complete, availability will also be available so we can filter out unavailable products without having to delete them entirely from the DB.